crm/pyproject.toml
[tool.poetry]
name = "flask-api-sample-crm"
version = "0.1.0"
description = "flask api sample crm using flask-restx framework"
authors = ["Aimene AHDIBI <ahdibiaymengmail.com>"]
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "^3.8.1"
flask = "^2.3.2"
flask-restx = "^1.1.0"
peewee = "^3.16.2"
uwsgi = "^2.0.21"
python-dotenv = "^1.0.0"
flask-jwt-extended = "^4.4.4"
pyscopg2 = "^66.0.2"
psycopg2-binary = "^2.9.6"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
flake8 = "^6.0.0"
isort = "^5.12.0"
pytest = "^7.3.1"
pre-commit = "^3.3.1"
coverage = "^7.2.5"
faker = "^18.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
